+++ src/value/sound.rs
@@ -11,7 +11,7 @@ impl PacketEncode for Sound {
     fn encode(&self, buf : &mut PacketBuf) -> Result<(), EncodeError> {
         match (self) {
             Sound::Registry(entry) => {
-                buf.encode_write(VarInt::from(entry.id() + 1))?;
+                buf.encode_write(VarInt::from((entry.id() as i32) + 1))?;
             },
             Sound::Inline(sound) => {
                 buf.encode_write(VarInt::from(0))?;
@@ -23,7 +23,7 @@ impl PacketEncode for Sound {
 }
 impl PacketDecode for Sound {
     fn decode(buf : &mut PacketBuf) -> Result<Self, DecodeError> {
-        let id = buf.read_decode::<VarInt>()?.as_i32() as usize;
+        let id = buf.read_decode::<VarInt>()?.as_i32() as u32;
         if (id == 0) {
             Ok(Self::Inline(buf.read_decode::<SoundEvent>()?))
         } else {
